backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.0
cryptography>=41.0.5
dnspython==2.7.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
pymongo==4.12.1
PyMySQL==1.1.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
flask_cors
python-dotenv
google-cloud-aiplatform>=1.36.0
google-auth>=2.16.0
Pillow>=11.0.0
sentence-transformers>=4.0.0
